// src/data/holders.generated.ts
export const TRACKED_TOKENS_RAW: RawTokenSeed[] = [
  {
    "symbol": "CASHCAT",
    "name": "Cash Cat",
    "address": "0x020bfC650A365f8BB26819deAAbF3E21291018b4",
    "totalSupply": 1000000000
  },
  {
    "symbol": "TENDIES",
    "name": "TENDIES",
    "address": "0x45242320DBB855EeA8Fd36804C6487E10E97FCF9",
    "totalSupply": 1000000000
  },
  {
    "symbol": "PONS",
    "name": "Pons",
    "address": "0x39dBED3a2bd333467115dE45665cC57F813C4571",
    "totalSupply": 1000000000
  },
  {
    "symbol": "INDEX",
    "name": "The Index",
    "address": "0x56910D4409F3a0C78C64DD8D0545FF0705389870",
    "totalSupply": 1000000000
  },
  {
    "symbol": "STONKBROKER",
    "name": "StonkBroker",
    "address": "0xe934e36A439C94017B64a3FecE66AF12099aBF50",
    "totalSupply": 2695364457
  },
  {
    "symbol": "VEX",
    "name": "ProjectVex",
    "address": "0x8Ff92566f2e81BDd68EDfAa8cde73942A723796b",
    "totalSupply": 1000000000
  },
  {
    "symbol": "JUGGERNAUT",
    "name": "The Juggernaut",
    "address": "0xD7321801CAae694090694Ff55A9323139F043B88",
    "totalSupply": 1000000000
  },
  {
    "symbol": "WISHBONE",
    "name": "WISHBONE",
    "address": "0x77581054581B9c525E7dd7a0155DE43867532d03",
    "totalSupply": 1000000000
  },
  {
    "symbol": "WOOD",
    "name": "Sherwood Protocol",
    "address": "0xF8BC08092C06dB6148114DCf82AF881F1085f92b",
    "totalSupply": 1000000000
  },
  {
    "symbol": "HOODRAT",
    "name": "Hoodrat",
    "address": "0x8e62F281f282686fCa6dCB39288069a93fC23F1c",
    "totalSupply": 1000000000
  },
  {
    "symbol": "ARROW",
    "name": "Arrow",
    "address": "0xf2915d1e3C1B0c769d0c756Ec43F1c1f6c99cD03",
    "totalSupply": 10000000
  },
  {
    "symbol": "YOLO",
    "name": "YOLO",
    "address": "0x62C71cd34a52c30d894419CBcc55Db2aFA8032eA",
    "totalSupply": 1000000000
  },
  {
    "symbol": "RAXOL",
    "name": "RAXOL",
    "address": "0xf44702b17d9abD53815F703e772F35E9c71A53af",
    "totalSupply": 1000000000
  },
  {
    "symbol": "KARMA",
    "name": "Karma by Virtuals",
    "address": "0xB47f4702DEB124cb4eB6286bE83c9d84277C6239",
    "totalSupply": 1000000000
  },
  {
    "symbol": "HMM",
    "name": "Thinking Cat",
    "address": "0x7FE995a80075dF3Dc8Ae11A9b82c7FE4202CD87f",
    "totalSupply": 1000000000
  }
];